# graphite-rest/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ArtifactListOut",
        "type": "object"
      },
      "ArtifactOut": {
        "properties": {
          "artifact_type": {
            "title": "Artifact Type",
            "type": "string"
          },
          "citations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GraphEvidenceOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "study_step_id": {
            "format": "uuid",
            "title": "Study Step Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "study_step_id",
          "artifact_type",
          "content",
          "created_at"
        ],
        "title": "ArtifactOut",
        "type": "object"
      },
      "Body_transcribe_course_audio_api_v1_courses__course_id__audio_transcriptions_post": {
        "properties": {
          "audio": {
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "type": "string"
          }
        },
        "required": [
          "audio"
        ],
        "title": "Body_transcribe_course_audio_api_v1_courses__course_id__audio_transcriptions_post",
        "type": "object"
      },
...
      "CourseCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CourseCreate",
        "type": "object"
      },
      "CourseListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CourseOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CourseListOut",
        "type": "object"
      },
      "CourseOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "CourseOut",
        "type": "object"
      },
...
      "GraphEvidenceOut": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "chunk_id",
          "document_id",
          "label",
          "excerpt"
        ],
        "title": "GraphEvidenceOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TranscriptionResponse": {
        "description": "\u00a710.4 voice input: the backend's job is only to produce editable text \u2014\nthe frontend decides what happens with it next.",
        "properties": {
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "language_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Code"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "TranscriptionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "graphite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/courses": {
      "get": {
        "operationId": "list_courses_api_v1_courses_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Courses",
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "create_course_api_v1_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/v1/courses/{course_id}": {
      "get": {
        "operationId": "get_course_api_v1_courses__course_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Course",
        "tags": [
          "courses"
        ]
      }
    },
    "/api/v1/courses/{course_id}/audio-transcriptions": {
      "post": {
        "operationId": "transcribe_course_audio_api_v1_courses__course_id__audio_transcriptions_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "description": "If true, also save the transcript as a text document on this course (status=UPLOADED, ready for the same ingestion pipeline as an uploaded .txt file). Interpretation beyond the literal design-doc.md \u00a710.4 spec, which only requires returning editable text to the caller.",
            "in": "query",
            "name": "persist_as_document",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true, also save the transcript as a text document on this course (status=UPLOADED, ready for the same ingestion pipeline as an uploaded .txt file). Interpretation beyond the literal design-doc.md \u00a710.4 spec, which only requires returning editable text to the caller.",
              "title": "Persist As Document",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_course_audio_api_v1_courses__course_id__audio_transcriptions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe Course Audio",
        "tags": [
          "audio"
        ]
      }
    },
...
    "/api/v1/study-steps/{step_id}/artifacts": {
      "get": {
        "operationId": "list_artifacts_api_v1_study_steps__step_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Step Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Artifacts",
        "tags": [
          "study-artifacts"
        ]
      },
      "post": {
        "description": "Generate (or return the cached) study material for one route step (\u00a710.1).",
        "operationId": "create_artifact_api_v1_study_steps__step_id__artifacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Step Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": true,
            "schema": {
              "enum": [
                "SUMMARY",
                "FLASHCARDS",
                "QUESTIONS"
              ],
              "title": "Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Artifact",
        "tags": [
          "study-artifacts"
        ]
      }
    },
...
  }
}